Move internal Ingress type from extensions to networking
