+++ pkg/apis/extensions/register.go
@@ -56,8 +56,8 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&ReplicationControllerDummy{},
 		&apps.DaemonSetList{},
 		&apps.DaemonSet{},
-		&Ingress{},
-		&IngressList{},
+		&networking.Ingress{},
+		&networking.IngressList{},
 		&apps.ReplicaSet{},
 		&apps.ReplicaSetList{},
 		&policy.PodSecurityPolicy{},
