+++ test/test_owners.csv
@@ -569,7 +569,6 @@ k8s.io/kubernetes/pkg/apis/batch/validation,erictune,0,
 k8s.io/kubernetes/pkg/apis/componentconfig,jbeda,1,
 k8s.io/kubernetes/pkg/apis/extensions,bgrant0607,1,
 k8s.io/kubernetes/pkg/apis/extensions/v1beta1,madhusudancs,1,
-k8s.io/kubernetes/pkg/apis/extensions/validation,nikhiljindal,1,
 k8s.io/kubernetes/pkg/apis/policy/validation,deads2k,1,
 k8s.io/kubernetes/pkg/apis/rbac/v1alpha1,liggitt,0,
 k8s.io/kubernetes/pkg/apis/rbac/validation,erictune,0,
@@ -745,8 +744,6 @@ k8s.io/kubernetes/pkg/registry/extensions/daemonset,nikhiljindal,1,
 k8s.io/kubernetes/pkg/registry/extensions/daemonset/storage,kevin-wangzefeng,1,
 k8s.io/kubernetes/pkg/registry/extensions/deployment,dchen1107,1,
 k8s.io/kubernetes/pkg/registry/extensions/deployment/storage,timothysc,1,
-k8s.io/kubernetes/pkg/registry/extensions/ingress,apelisse,1,
-k8s.io/kubernetes/pkg/registry/extensions/ingress/storage,luxas,1,
 k8s.io/kubernetes/pkg/registry/extensions/replicaset,rrati,0,
 k8s.io/kubernetes/pkg/registry/extensions/replicaset/storage,wojtek-t,1,
 k8s.io/kubernetes/pkg/registry/extensions/rest,rrati,0,
